requirements.txt
python-telegram-bot>=20.8,<21
discord.py>=2.3,<3
python-dotenv>=1.0,<2
openai>=1.51.0
